+++ pccl/include/common_hip.h
@@ -16,9 +16,9 @@
 } while(0)
 
 // Kernel for vector addition.
-__global__ void vectorAddKernel(float* a, const float* b, int n);
+__global__ void vectorAddKernel(float* a, const float* b, int64_t n);
 
 // Function to launch the kernel.
-void vectorAdd(float* a, const float* b, int n, hipStream_t stream);
+void vectorAdd(float* a, const float* b, int64_t n, hipStream_t stream);
 
 #endif // COMMON_H
